requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
astroid==2.4.2
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
flask-restplus==0.13.0
Flask-Sijax==0.4.1
Flask-SQLAlchemy==2.4.4
future==0.18.2
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
numpy==1.19.5
pandas==1.2.0
Pillow==8.1.0
pipupgrade==1.9.0
psycopg2==2.8.6
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
Sijax==0.3.2
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
urllib3==1.26.3
virtualenv==20.4.2
Werkzeug==0.16.0
wrapt==1.12.1
